+++ app/containers/AudioOutputMenu/index.js
@@ -2,6 +2,8 @@ import React, { Fragment } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 
+import { isToolbar } from 'utils/url';
+import { selectCrmModule } from 'containers/AgentDesktop/selectors';
 import { selectAudioPreferences } from 'containers/AgentNotificationsMenu/selectors';
 import {
   selectActiveOutputRingtoneDevices,
@@ -28,18 +30,18 @@ import DeviceMenu from './DeviceMenu';
  *    A Map containing the MediaDeviceInfo object of all available input devices (hardware devices
  *    capable of providing an input audio stream), indexed by deviceId.
  *
- * Media:
- *    https://www.twilio.com/docs/voice/client/javascript/device#twiliodeviceaudioringtonedevices
- *
- *    ringtoneDevices is an AudioOutputCollection that controls which output devices are used to
- *    play the ringing sound when receiving an incoming call.
- *
- * Notification:
+ * Media/Notification:
  *    https://developer.mozilla.org/en-US/docs/Web/API/HTMLAudioElement/Audio
  *
  *    The Audio() constructor creates and returns a new HTMLAudioElement which can be either
  *    attached to a document for the user to interact with and/or listen to,
  *    or can be used offscreen to manage and play audio.
+ *
+ * Ringtone:
+ *    https://www.twilio.com/docs/voice/client/javascript/device#twiliodeviceaudioringtonedevices
+ *
+ *    ringtoneDevices is an AudioOutputCollection that controls which output devices are used to
+ *    play the ringing sound when receiving an incoming call.
  *
  * voice:
  *    https://www.twilio.com/docs/voice/client/javascript/device#twiliodeviceaudiospeakerdevices
@@ -50,21 +52,23 @@ import DeviceMenu from './DeviceMenu';
 
 export const AudioOutputMenu = (props) => (
   <Fragment>
-    <DeviceMenu
-      audio="media"
-      label={messages.media}
-      devices={props.ringtoneDevices}
-      setDeviceAsActive={props.updateActiveOutputRingtoneDevice}
-    />
-
-    {props.audioNotificationsEnabled && (
+    {(props.audioNotificationsEnabled ||
+      !isToolbar() ||
+      props.crmModule === 'zendesk') && (
       <DeviceMenu
-        audio="notifications"
-        label={messages.notifications}
+        audio="media"
+        label={messages.media}
         devices={props.notificationDevices}
         setDeviceAsActive={props.updateActiveOutputNotificationDevice}
       />
     )}
+
+    <DeviceMenu
+      audio="ringtone"
+      label={messages.ringtone}
+      devices={props.ringtoneDevices}
+      setDeviceAsActive={props.updateActiveOutputRingtoneDevice}
+    />
 
     <DeviceMenu
       audio="voice"
@@ -80,6 +84,7 @@ const mapStateToProps = (state, props) => ({
   speakerDevices: selectActiveOutputSpeakerDevices(state, props),
   notificationDevices: selectActiveOutputNotificationDevices(state, props),
   audioNotificationsEnabled: selectAudioPreferences(state, props),
+  crmModule: selectCrmModule(state, props),
 });
 
 const actions = {
@@ -111,6 +116,7 @@ AudioOutputMenu.propTypes = {
   updateActiveOutputRingtoneDevice: PropTypes.func.isRequired,
   updateActiveOutputSpeakerDevice: PropTypes.func.isRequired,
   updateActiveOutputNotificationDevice: PropTypes.func.isRequired,
+  crmModule: PropTypes.string,
 };
 
 export default connect(